moved compare_spectra in the compare module
