+++ radis/test/spectrum/test_spectrum.py
@@ -166,6 +166,7 @@ def test_store_functions(verbose=True, *args, **kwargs):
     temp_file = 'test_radis_tempfile_transmittance.txt'
     assert not exists(temp_file)
 
+    # Test that the transmittance stored on .txt and loaded again match
     s = load_spec(getTestFile(
         'CO_Tgas1500K_mole_fraction0.01.spec'), binary=True)
     s.update()
@@ -179,6 +180,9 @@ def test_store_functions(verbose=True, *args, **kwargs):
     s2 = transmittance_spectrum(
         w, T, wunit='nm', conditions={'medium': 'vacuum'})
     assert s.compare_with(s2, spectra_only='transmittance_noslit', plot=False)
+
+    # TODO: add test that ensures we can load a binary file without binary=True
+    # (and the warning should be captured)
 
     return True
